docker-isys/requirement.txt
adal==1.2.2
aiohttp==3.5.4
alabaster==0.7.12
alembic==0.9.10
amqp==2.5.0
analytics-python==1.2.9
apache-airflow==1.10.3
asn1crypto==0.24.0
astroid==2.2.5
async-timeout==3.0.1
atlasclient==0.1.6
attrs==19.1.0
aws-xray-sdk==0.95
azure-common==1.1.23
azure-cosmos==3.1.0
azure-datalake-store==0.0.19
azure-mgmt-containerinstance==1.5.0
azure-mgmt-datalake-nspkg==3.0.1
azure-mgmt-datalake-store==0.4.0
azure-mgmt-nspkg==3.0.2
azure-mgmt-resource==1.2.2
azure-nspkg==3.0.2
azure-storage==0.36.0
azure-storage-blob==2.0.1
azure-storage-common==2.0.0
Babel==2.7.0
bcrypt==3.1.7
beautifulsoup4==4.7.1
billiard==3.5.0.5
boto==2.49.0
boto3==1.7.84
botocore==1.10.84
cachetools==3.1.1
cassandra-driver==3.18.0
celery==4.1.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
cloudant==0.5.10
colorama==0.4.1
configparser==3.5.3
cookies==2.2.1
croniter==0.3.30
cryptography==2.7
cx-Oracle==7.2.0
datadog==0.29.3
decorator==4.4.0
defusedxml==0.6.0
dill==0.2.9
docker==3.7.3
docker-pycreds==0.4.0
docopt==0.6.2
docutils==0.14
ecdsa==0.13.2
elasticsearch==5.5.3
elasticsearch-dsl==5.4.0
entrypoints==0.3
fastavro==0.22.2
flake8==3.7.7
Flask==1.0.3
Flask-Admin==1.5.3
Flask-AppBuilder==1.12.3
Flask-Babel==0.12.2
Flask-BabelEx==0.9.3
Flask-Bcrypt==0.7.1
Flask-Caching==1.3.3
Flask-Login==0.4.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.4.0
flask-swagger==0.2.13
Flask-WTF==0.14.2
flower==0.9.3
freezegun==0.3.12
funcsigs==1.0.0
future==0.16.0
future-fstrings==1.2.0
gitdb2==2.0.5
GitPython==2.1.11
google-api-core==1.13.0
google-api-python-client==1.7.9
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
google-cloud-bigquery==1.16.0
google-cloud-bigtable==0.31.0
google-cloud-container==0.2.1
google-cloud-core==0.28.1
google-cloud-language==1.2.0
google-cloud-spanner==1.9.0
google-cloud-translate==1.5.0
google-cloud-vision==0.37.0
google-resumable-media==0.3.2
googleapis-common-protos==1.6.0
grpc-google-iam-v1==0.11.4
grpcio==1.22.0
grpcio-gcp==0.2.2
gunicorn==19.9.0
hdfs==2.5.8
hmsclient==0.1.1
httplib2==0.9.2
idna==2.8
ijson==2.4
imagesize==1.1.0
inflection==0.3.1
iso8601==0.1.12
isodate==0.6.0
itsdangerous==1.1.0
JayDeBeApi==1.1.1
Jinja2==2.10
jira==2.0.0
jmespath==0.9.4
JPype1==0.7.0
json-merge-patch==0.2
jsondiff==1.1.1
jsonpickle==1.2
kombu==4.6.3
kubernetes==10.0.0
lazy-object-proxy==1.4.1
ldap3==2.6
lockfile==0.12.2
lxml==4.3.4
Mako==1.0.13
Markdown==2.6.11
MarkupSafe==1.1.1
mccabe==0.6.1
mock==3.0.5
mongomock==3.17.0
moto==1.3.5
msrest==0.4.29
msrestazure==0.6.1
multi-key-dict==2.0.3
multidict==4.5.2
mysqlclient==1.3.14
nose==1.3.7
nose-ignore-docstring==0.2
nose-timer==0.7.5
ntlm-auth==1.3.0
numpy==1.16.4
oauthlib==3.0.2
ordereddict==1.1
packaging==19.0
pandas==0.24.2
pandas-gbq==0.10.0
parameterized==0.7.0
paramiko==2.6.0
pbr==5.4.0
pendulum==1.4.4
pinotdb==0.1.1
protobuf==3.8.0
psutil==5.6.3
psycopg2==2.7.7
pyaml==19.4.1
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycodestyle==2.5.0
pycparser==2.19
pycryptodome==3.8.2
pycryptodomex==3.8.2
pydata-google-auth==0.1.3
pydruid==0.5.6
pyflakes==2.1.1
Pygments==2.4.2
PyHive==0.6.1
PyMySQL==0.9.3
PyJWT==1.7.1
pykerberos==1.2.1
pymongo==3.8.0
pymssql==2.1.4
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.0
pysftp==0.2.9
PySmbClient==0.1.5
python-daemon==2.1.2
python-dateutil==2.8.0
python-editor==1.0.4
python-http-client==3.1.0
python-jenkins==1.4.0
python-jose==2.0.2
python3-openid==3.1.0
pytz==2019.1
pytzdata==2019.2
pywinrm==0.3.0
PyYAML==5.1.1
qds-sdk==1.12.0
redis==3.2.1
rednose==1.3.0
requests==2.22.0
requests-futures==0.9.4
requests-kerberos==0.12.0
requests-mock==1.6.0
requests-ntlm==1.1.0
requests-oauthlib==1.2.0
requests-toolbelt==0.9.1
responses==0.10.6
rsa==4.0
s3transfer==0.1.13
sasl==0.2.1
sendgrid==5.6.0
sentinels==1.0.0
setproctitle==1.1.10
six==1.12.0
slackclient==2.1.0
smmap2==2.0.5
snakebite==2.11.0
snowballstemmer==1.9.0
snowflake-connector-python==1.8.4
snowflake-sqlalchemy==1.1.13
soupsieve==1.9.2
speaklater==1.3
Sphinx==2.1.2
sphinx-argparse==0.2.5
sphinx-autoapi==1.1.0
Sphinx-PyPI-upload==0.2.1
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-dotnetdomain==0.4
sphinxcontrib-golangdomain==0.2.0.dev0
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-httpdomain==1.7.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
SQLAlchemy==1.2.19
sshtunnel==0.1.5
tabulate==0.8.3
tenacity==4.12.0
termstyle==0.1.11
text-unidecode==1.2
thrift==0.11.0
thrift-sasl==0.3.0
tornado==5.1.1
typed-ast==1.4.0
tzlocal==1.5.1
unicodecsv==0.14.1
Unidecode==1.1.1
uritemplate==3.0.0
urllib3==1.25.3
vertica-python==0.9.3
vine==1.3.0
websocket-client==0.56.0
Werkzeug==0.14.1
wrapt==1.11.2
WTForms==2.2.1
xlrd==1.2.0
xmltodict==0.12.0
yarl==1.3.0
zdesk==2.7.1
zope.deprecation==4.4.0
